updating according to feedback
