+++ jenkins-build-pack-lambda/Jenkinsfile
@@ -43,7 +43,8 @@ def Event_Status = [
 	'FAILED':'FAILED'
 ]
 
-@Field def envLoader
+@Field def configModule
+@Field def configLoader
 @Field def var_api_key
 @Field def var_api_id_dev
 @Field def var_aws_credential_id
@@ -66,20 +67,20 @@ def Event_Status = [
 @Field def jq = '/usr/local/bin/jq'
 
 node ()  {
-	loadEnvModule("http://${env.REPO_BASE}/scm/${env.REPO_CORE}/jazz-build-module.git")
-	var_api_key 				= envLoader.getApiField("API_KEY")
-	var_api_id_dev = envLoader.getApiField("API_ID_DEV")
-	var_aws_credential_id = envLoader.getAWSCredentialID()
-	var_env_name_prefix = envLoader.getEnvNamePrefix()
-	var_repo_core = envLoader.getRepositoryField("REPO_CORE")
-	var_repo_loc = envLoader.getRepositoryField("REPO_LOC")
-	var_credentialsId 			= envLoader.getRepositoryField("REPO_CREDENTIAL_ID")
-	var_bitbucket_base			= envLoader.getRepositoryField("REPO_BASE")
-	var_bitbucket_email 		= envLoader.getRepositoryField("REPO_EMAIL")
-	region = envLoader.getJazzField("jazz_region")
-	g_base_url = "https://${envLoader.getApiField("API_KEY")}.execute-api.${region}.amazonaws.com/dev"
-	role = envLoader.getJazzField("jazz_roleId")
-	api_doc_s3_bucket = envLoader.getJazzS3Field("jazz_s3_api_doc")
+	loadConfigModule("http://${env.REPO_BASE}/scm/${env.REPO_CORE}/jazz-build-module.git")
+	var_api_key 				= configLoader.API.API_KEY
+	var_api_id_dev = configLoader.API.API_ID_DEV
+	var_aws_credential_id = configLoader.AWS_CREDENTIAL_ID
+	var_env_name_prefix = configLoader.env_name_prefix
+	var_repo_core = configLoader.REPOSITORY.REPO_CORE
+	var_repo_loc = configLoader.REPOSITORY.REPO_LOC
+	var_credentialsId 			= configLoader.REPOSITORY.REPO_CREDENTIAL_ID
+	var_bitbucket_base			= configLoader.REPOSITORY.REPO_BASE
+	var_bitbucket_email 		= configLoader.REPOSITORY.REPO_EMAIL
+	region = configLoader.JAZZ.jazz_region
+	g_base_url = "https://${configLoader.API.API_KEY}.execute-api.${region}.amazonaws.com/dev"
+	role = configLoader.JAZZ.jazz_roleId
+	api_doc_s3_bucket = configLoader.JAZZ_S3.jazz_s3_api_doc
 
 	echo "Jobs Params: " + params
 	echo "Build triggered via branch: " + params.scm_branch
@@ -217,7 +218,7 @@ node ()  {
 						setLambdaExecutionRole(role);
 						sendEvent(service, branch, runtime, "PRODUCTION", region, domain, role, _event, Event_Status.COMPLETED, "")
 
-						sh "serverless deploy --stage prod -v --bucket " + envLoader.getJazzS3Field("jazz_bucket_prod")
+						sh "serverless deploy --stage prod -v --bucket " + configLoader.JAZZ_S3.jazz_bucket_prod
 
 						def stackProdName = cfStackName + "-prod"
 						lambdaProdARN = getLambdaARN(stackProdName)
@@ -283,7 +284,7 @@ node ()  {
 						setLambdaExecutionRole(role);
 						sendEvent(service, branch, runtime, "DEVELOPMENT", region, domain, role, _event, Event_Status.COMPLETED, "")
 
-						sh "serverless deploy --stage dev  -v --bucket " + envLoader.getJazzS3Field("jazz_bucket_dev")
+						sh "serverless deploy --stage dev  -v --bucket " + configLoader.JAZZ_S3.jazz_bucket_dev
 						def stackDevName = cfStackName + "-dev";
 						lambdaDevARN = getLambdaARN(stackDevName);
 						echo "lambdaDevARN: ${lambdaDevARN}"
@@ -806,10 +807,10 @@ def updateServiceInDB(service_id, domain, devApiID, region, status) {
 /*
 * Load environment variables from build module
 */
-def loadEnvModule(buildModuleUrl){
+def loadConfigModule(buildModuleUrl){
 	echo "loading env variables, checking repos..."
 
-	dir('env-loader') {
+	dir('config-loader') {
 		checkout([$class: 'GitSCM', branches: [
 			[name: '*/master']
 		], doGenerateSubmoduleConfigurations: false, extensions: [], submoduleCfg: [], userRemoteConfigs: [
@@ -819,8 +820,8 @@ def loadEnvModule(buildModuleUrl){
 		echo "loading installer variables..."
 
 		def resultJsonString = readFile("jazz-installer-vars.json")
-		envLoader = load "env-loader.groovy"
-		envLoader.initialize(resultJsonString)
+		configModule = load "config-loader.groovy"
+		configLoader = configModule.initialize(resultJsonString)
 		echo "finished loading env module"
 	}
 }
